+++ app/views/projects/show.html.erb
@@ -23,11 +23,20 @@
           <strong><%= link_to "Voir plus", company_path(@chosen_company), class: "text-decoration-none" %></strong>
         </small></li>
       </ul>
-
     </div>
+
     <div id="project-virtual-visit">
       <h3><i class="fa-solid fa-eye me-2"></i> Ma visite virtuelle</h3>
       prévue le <strong><%= @project.created_at.strftime("%d/%m/%Y")%></strong>
+      <div data-controller="video-call"
+        data-video-call-access-token-value="<%= @twilio_jwt %>"
+        data-video-call-room-id-value="<%= @room_id %>"
+        data-video-call-buddy-video-width-value="200">
+        <div data-video-call-target="buddyVideo"></div>
+        <div data-video-call-target="localVideo"></div>
+        <div class="btn-call-start" data-action="click->video-call#joinCall">Call</div>
+        <div class="btn-call-end" data-action="click->video-call#endCall">End call</div>
+      </div>
     </div>
   </div>
